Added [ContributedAction] -  see commit comment.
I have added this because I think we are going to need it unless (or until) the functions for 'object actions' and 'main menu actions' are registered via separate mechanisms.  In which case we could just adopt the convention that for functions registered as object actions the first paramater must be a 'functional type' and will be taken as the contributee. This would be cleaner than [ContributedAction] certainly.